% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{cholesky_inverse}
\alias{cholesky_inverse}
\title{Compute an inverse using the Cholesky decomposition}
\usage{
cholesky_inverse(matrix_to_invert)
}
\arguments{
\item{matrix_to_invert}{The matrix to find the inverse for.}
}
\value{
The inverse of the matrix given.
}
\description{
This function computes the inverse of a positive definite matrix using the
Cholesky decomposition.
}
