% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{diag_rbf_kernel}
\alias{diag_rbf_kernel}
\title{Compute a diagonal RBF kernel}
\usage{
diag_rbf_kernel(v1, v2, l, tau)
}
\arguments{
\item{v1}{The first vector to compute distances for.}

\item{v2}{The second vector to compute distances for.}

\item{l}{The length scale of the RBF kernel.}

\item{tau}{The standard deviation of the kernel.}
}
\value{
A vector containing the diagonal elements of the RBF kernel matrix
between v1 and v2.
}
\description{
This function computes the diagonal RBF kernel for two vectors.
}
