% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{fit_marginal_likelihood_rbf}
\alias{fit_marginal_likelihood_rbf}
\title{Fits the hyperparameters of a single RBF kernel by mAP estimation}
\usage{
fit_marginal_likelihood_rbf(x_train, y_train, start_sigma = 10,
  start_l = 10, start_tau = 10)
}
\arguments{
\item{x_train}{The training covariates.}

\item{y_train}{The training observations.}

\item{start_sigma}{The initial observation noise to use.}

\item{start_l}{The initial length scale to use.}

\item{start_tau}{The initial process standard deviation to use.}
}
\value{
A list containing entries `sigma`, `l` and `tau` -- the optimised
hyperparameters of the RBF kernel.
}
\description{
This function optimises the hyperparameters, as well as the observation
noise, for an RBF kernel.
}
