% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{full_rbf_kernel}
\alias{full_rbf_kernel}
\title{Compute an RBF Kernel}
\usage{
full_rbf_kernel(v1, v2, l, tau)
}
\arguments{
\item{v1}{The first vector to compute distances for.}

\item{v2}{The second vector to compute distances for.}

\item{l}{The length scale of the RBF kernel.}

\item{tau}{The standard deviation of the kernel.}
}
\value{
A matrix containing the distances between each of the elements in v1
and v2.
}
\description{
This function computes the full RBF kernel for two vectors.
}
