% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_functions.R
\name{predict_points}
\alias{predict_points}
\title{Fit an exact GP and predict a set of points}
\usage{
predict_points(x_train, x_new, sigma_noise, y, kernel_fun, mean_centre = TRUE)
}
\arguments{
\item{x_train}{The training covariates.}

\item{x_new}{The covariates to use for prediction,}

\item{sigma_noise}{The observation noise.}

\item{y}{The training observations.}

\item{kernel_fun}{The kernel to use.}
}
\value{
A list containing entries `mean` and `cov` specifying the mean and
covariance at the points given by `x_new`.
}
\description{
This function fits and predicts a GP using the exact expressions for mean and
covariance.
}
